chore: rename from ignored directory
