+++ src/index.js
@@ -4,7 +4,7 @@ import { Bitswap } from './bitswap.js'
  * @typedef {import('./types').IPFSBitswap} IPFSBitswap
  * @typedef {import('./types').MultihashHasherLoader} MultihashHasherLoader
  * @typedef {import('peer-id')} PeerId
- * @typedef {import('./types/message')} BitswapMessage
+ * @typedef {import('./message')} BitswapMessage
  * @typedef {import('interface-blockstore').Blockstore} Blockstore
  * @typedef {import('interface-blockstore').Pair} Pair
  * @typedef {import('interface-blockstore').Options} Options
