+++ src/wantlist/entry.js
