+++ test/bitswap-mock-internals.js
@@ -5,7 +5,7 @@ import { expect } from 'aegir/utils/chai.js'
 import PeerId from 'peer-id'
 import all from 'it-all'
 import drain from 'it-drain'
-import { BitswapMessage as Message } from '../src/types/message/index.js'
+import { BitswapMessage as Message } from '../src/message/index.js'
 import { Bitswap } from '../src/bitswap.js'
 import { CID } from 'multiformats/cid'
 import { AbortController } from 'native-abort-controller'
