+++ test/network/network.node.js
@@ -7,7 +7,7 @@ import pDefer from 'p-defer'
 import { createLibp2pNode } from '../utils/create-libp2p-node.js'
 import { makeBlocks } from '../utils/make-blocks.js'
 import { Network } from '../../src/network.js'
-import { BitswapMessage as Message } from '../../src/types/message/index.js'
+import { BitswapMessage as Message } from '../../src/message/index.js'
 import { Stats } from '../../src/stats/index.js'
 import sinon from 'sinon'
 import { CID } from 'multiformats/cid'
