+++ test/types/message.spec.js
@@ -9,8 +9,8 @@ import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { concat as uint8ArrayConcat } from 'uint8arrays/concat'
 import loadFixture from 'aegir/utils/fixtures.js'
 import varint from 'varint'
-import { Message } from '../../src/types/message/message.js'
-import { BitswapMessage } from '../../src/types/message/index.js'
+import { Message } from '../../src/message/message.js'
+import { BitswapMessage } from '../../src/message/index.js'
 import { makeBlocks } from '../utils/make-blocks.js'
 
 const testDataPath = 'test/fixtures/serialized-from-go'
