+++ tsconfig.json
@@ -8,7 +8,7 @@
     "test"
   ],
   "exclude": [
-    "src/types/message/message.js", // generated file
+    "src/message/message.js", // generated file
     "test/types.test-d.ts"
   ]
 }
